Fix in finding geometrically close tracksters for linking
Not every trackster in the tiles would be within delta of the seed,
this is fixed by an additional check on the tracksters returned.

Tracksters found are then sorted in the ascending order of the their
distances from the seed. This way, when exploring the links found,
the ones that are closest would be considered first.

Minor changes in comments, energy regression and ID for merged
tracksters.